Making a Pod out of NYBlurView

+++ Pod/UIImage+ImageEffects.m
@@ -223,7 +223,7 @@
 
         if (hasBlur) {
             CGFloat inputRadius = blurRadius * 1;
-            NSUInteger radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
+            unsigned int radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
             
             if (radius % 2 != 1) {
                 radius += 1;
